+++ Domain/Entities/Transaction.cs
@@ -4,7 +4,7 @@
     {
         public int WalletId { get; set; }
         public string TransactionType { get; set; }
-        public decimal Amount { get; set; }
+        public Int64 Amount { get; set; }
         public string? Description { get; set; }
         public DateTime TransactionDate { get; set; }
         public string Status { get; set; }
