+++ Services/DTO/EventOrderDTOs/EventOrderDetailsReponseDTO.cs
@@ -6,6 +6,6 @@
         public int PackageId { get; set; }
         public int EventOrderId { get; set; }
         public int Quantity { get; set; }
-        public decimal Price { get; set; }
+        public Int64 Price { get; set; }
     }
 }
