+++ Services/DTO/WalletDTOs/WalletResponseDTO.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public int UserId { get; set; }
         public string WalletType { get; set; }
-        public decimal Balance { get; set; }
+        public Int64 Balance { get; set; }
     }
 }
