+++ Services/Interface/IWalletService.cs
@@ -7,7 +7,7 @@ namespace Services.Interface
     public interface IWalletService
     {
         Task<TransactionResponsesDTO> ConfirmTransaction(int transactionId);
-        Task<Transaction> Deposit(int userId, decimal amount);
+        Task<Transaction> Deposit(int userId, Int64 amount);
         Task<List<WalletResponseDTO>> GetListWalletByUserId(int userId);
         Task<Transaction> GetTransactionById(int transactionId);
         Task<List<TransactionResponsesDTO>> GetTransactions(int userId);
